Better tests

+++ src/test/java/org/thshsh/sas/TestFile.java
@@ -4,8 +4,9 @@ import java.io.File;
 import java.net.URI;
 
 public class TestFile {
+	
 	public File file;
-	public Integer rows;
+	public Long rows;
 	public Integer columns;
 	public TestFile(URI uri) {
 		this(new File(uri));
@@ -17,11 +18,11 @@ public class TestFile {
 			String[] size = parts[parts.length-1].split("x");
 			if(size.length == 2) {
 				this.columns = Integer.valueOf(size[0]);
-				this.rows = Integer.valueOf(size[1]);
+				this.rows = Long.valueOf(size[1]);
 			}
 		}
 	}
-	public TestFile(URI uri, Integer rows, Integer columns) {
+	public TestFile(URI uri, Long rows, Integer columns) {
 		super();
 		this.file = new File(uri);
 		this.rows = rows;
